requirements.txt
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.3
spotipy==2.23.0
gunicorn==21.2.0
python-dotenv